+++ addOns/insights/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Correct Automation Framework `insights` job template loading.
 
 ## [0.5.0] - 2026-06-12
 ### Added